xong phần tạo video

+++ hoc-laravel/app/Http/Controllers/clients/VideoController.php
@@ -46,13 +46,24 @@ class VideoController extends Controller{
     public function edit(Request $request) {
         $data = request()->all();
 
+        $userId = session('loggedInUser');
+
         $video_id = $data['video_id'];
 
         $video = Video::find($video_id);
 
+        $newVideoPath = null;
+        $newThumbnailPath = null;
+
+        //nếu video tồn tại thì lấy đường dẫn video và thumbnail cũ
+        if($video){
+            $newVideoPath = $video->video_path;
+            $newThumbnailPath = $video->thumbnail_path;
+        }
+
         if($request->hasFile('video_path')){
             //xóa video cũ
-            if($video->video_path){
+            if($video && $video->video_path){
                 Storage::delete('public/video/' . $video->video_path);
             }
             //lấy file từ request
@@ -60,9 +71,9 @@ class VideoController extends Controller{
             //lấy extension của file
             $videoExtension = $video_file->getClientOriginalExtension();
             //tạo tên mới cho video
-            $video_path_name =  $video->video_id . time() . '.' . $videoExtension;
+            $video_path_name =  time() . '.' . $videoExtension;
             //lưu đường dẫn video mới
-            $video->video_path = $video_path_name;
+            $newVideoPath = $video_path_name;
             //lưu video
             $video_file->storeAs('public/video/', $video_path_name);
         }
@@ -70,7 +81,7 @@ class VideoController extends Controller{
         if($request->hasFile('thumbnail_path')){
 
             //xóa thumbnail cũ
-            if($video->thumbnail_path){
+            if($video && $video->thumbnail_path){
                 Storage::delete('public/thumbnail/' . $video->thumbnail_path);
             }
             //lấy file từ request
@@ -78,9 +89,9 @@ class VideoController extends Controller{
             //lấy extension của file
             $thumbnailExtension = $thumbnail_file->getClientOriginalExtension();
             //tạo tên mới cho thumbnail
-            $thumbnail_path_name =  $video->video_id . time() . '.' . $thumbnailExtension;
+            $thumbnail_path_name =  time() . '.' . $thumbnailExtension;
             //lưu đường dẫn thumbnail mới
-            $video->thumbnail_path = $thumbnail_path_name;
+            $newThumbnailPath = $thumbnail_path_name;
             //lưu thumbnail
             $thumbnail_file->storeAs('public/thumbnail/', $thumbnail_path_name);
         }
@@ -88,14 +99,30 @@ class VideoController extends Controller{
         $newVideo = [
             'video_id' => $data['video_id'],
             'title' => $data['title'],
+            'user_id' => $userId,
             'description' => $data['description'],
             'display_mode' => $data['display_mode'],
-            'video_path' => $video->video_path,
-            'thumbnail_path' => $video->thumbnail_path,
+            'video_path' => $newVideoPath,
+            'thumbnail_path' => $newThumbnailPath,
         ];
 
         $videoModel = new Video();
-        $videoModel->updateVideo($video_id, $newVideo);
+
+        if($data['video_id']){
+            $videoModel->updateVideo($video_id, $newVideo);
+        }else{
+            unset($newVideo['video_id']);
+            $newVideo['created_date'] = date('Y-m-d H:i:s');
+            $newVideo['active'] = 1;
+            $newVideo['view'] = 0;
+
+            //cái này sẽ cần sửa lại thêm trong giao diện
+            $newVideo['membership'] = 0;
+
+            //khi nào xong chức năng duyệt video thì sửa lại thành 0
+            $newVideo['is_approved'] = 1;
+            $videoModel->createVideo($newVideo);
+        }
 
         return response()->json($newVideo);
     }
